apply workflow not before entity change for proper create notifications

+++ bundles/org.zikula.modulestudio.generator/src/org/zikula/modulestudio/generator/cartridges/zclassic/controller/helper/WorkflowHelper.xtend
@@ -421,16 +421,19 @@ class WorkflowHelper {
                 }
             «ENDIF»
             $result = false;
+            if (!$workflow->can($entity, $actionId)) {
+                return $result;
+            }
 
             try {
-                $workflow->apply($entity, $actionId);
-
                 if ('delete' == $actionId) {
                     $entityManager->remove($entity);
                 } else {
                     $entityManager->persist($entity);
                 }
                 $entityManager->flush();
+
+                $workflow->apply($entity, $actionId);
 
                 $result = true;
                 if ('delete' == $actionId) {
